Moved towards fixing broken code in master branch.

+++ webapp/app/forms.py
@@ -1,6 +1,8 @@
 from flask_wtf import Form
-from wtforms import TextField, TextAreaField, validators, PasswordField, StringField, SubmitField, FloatField, EmailField
-from wtforms.validators import DataRequired, EqualTo, NumberRange
+from wtforms import validators
+from wtforms.validators import DataRequired, EqualTo, NumberRange, Email
+from wtforms.fields.html5 import EmailField
+from wtforms.fields import TextField, TextAreaField, PasswordField, StringField, SubmitField, DateField, IntegerField
 
 class CreateAccountForm(Form):
     email = EmailField('Email Address', [validators.DataRequired(), validators.Email()])
@@ -8,10 +10,10 @@ class CreateAccountForm(Form):
     forename = StringField('Forename', validators=[DataRequired()])
     surname = StringField('Surname', validators=[DataRequired()])
     date_of_birth = DateField('Date of Birth', validators=[DataRequired()])
-    card_number = FloatField('Card Number', [validators.DataRequired(), validators.NumberRange(min=0, max=9999999999999999)])
-    cvc = FloatField('CVC', [validators.DataRequired(), validators.NumberRange(min=0, max=999)])
+    card_number = IntegerField('Card Number', [validators.DataRequired(), validators.NumberRange(min=0, max=9999999999999999)])
+    cvc = IntegerField('CVC', [validators.DataRequired(), validators.NumberRange(min=0, max=999)])
     expiry_date_month = DateField('Expiry Date', [validators.DataRequired, validators.NumberRange(min=1, max=12)])
-    expiry_date_year = FloatField('Expiry Date Year' [validators.DataRequired, validators.NumberRange(min=2018)])
+    expiry_date_year = IntegerField('Expiry Date Year', [validators.DataRequired, validators.NumberRange(min=2018)])
     submit = SubmitField('Register')
 
 class ChangePasswordForm(Form):
